Fix existing cancellation tests
