+++ posthog/api/test/batch_exports/test_runs.py
@@ -1,8 +1,8 @@
 import asyncio
+import datetime as dt
+from unittest.mock import patch
 
 import pytest
-import temporalio.client
-from asgiref.sync import async_to_sync
 from django.test.client import Client as HttpClient
 from rest_framework import status
 
@@ -15,9 +15,11 @@ from posthog.api.test.batch_exports.operations import (
     get_batch_export,
     get_batch_export_runs,
     get_batch_export_runs_ok,
+    wait_for_workflow_executions,
 )
 from posthog.api.test.test_team import create_team
 from posthog.api.test.test_user import create_user
+from posthog.test.base import _create_event
 
 pytestmark = [
     pytest.mark.django_db,
@@ -145,28 +147,6 @@ def test_batch_exports_are_partitioned_by_team(client: HttpClient, temporal):
         assert response.status_code == status.HTTP_404_NOT_FOUND, response.json()
 
 
-# TODO - this was in test_delete.py too so maybe extract it out into operations.py?
-@async_to_sync
-async def wait_for_workflow_executions(
-    temporal: temporalio.client.Client, query: str, timeout: int = 30, sleep: int = 1
-):
-    """Wait for Workflow Executions matching query."""
-    workflows = [workflow async for workflow in temporal.list_workflows(query=query)]
-
-    total = 0
-    while not workflows:
-        total += sleep
-
-        if total > timeout:
-            raise TimeoutError(f"No backfill Workflow Executions after {timeout} seconds")
-
-        await asyncio.sleep(sleep)
-        workflows = [workflow async for workflow in temporal.list_workflows(query=query)]
-
-    return workflows
-
-
-@pytest.mark.skip("Flaky test failing")
 @pytest.mark.django_db(transaction=True)
 def test_cancelling_a_batch_export_run(client: HttpClient, temporal):
     """Test cancelling a BatchExportRun."""
@@ -191,33 +171,48 @@ def test_cancelling_a_batch_export_run(client: HttpClient, temporal):
     user = create_user("test@user.com", "Test User", organization)
     client.force_login(user)
 
-    with start_test_worker(temporal):
-        batch_export = create_batch_export_ok(
-            client,
-            team.pk,
-            batch_export_data,
-        )
-        batch_export_id = batch_export["id"]
+    with patch("posthog.temporal.batch_exports.s3_batch_export.Producer.start") as mock_producer_start:
+        # Mock the producer to sleep so we can test cancellation
+        async def mock_producer_start(*args, **kwargs):
+            await asyncio.sleep(5)
+            return None
 
-        start_at = "2023-10-23T00:00:00+00:00"
-        end_at = "2023-10-24T00:00:00+00:00"
-        backfill_batch_export_ok(client, team.pk, batch_export_id, start_at, end_at)
+        mock_producer_start.side_effect = mock_producer_start
 
-        # In order for a run to be cancelable we need a running workflow execution
-        _ = wait_for_workflow_executions(temporal, query=f'TemporalScheduledById="{batch_export_id}"')
+        with start_test_worker(temporal):
+            batch_export = create_batch_export_ok(
+                client,
+                team.pk,
+                batch_export_data,
+            )
+            batch_export_id = batch_export["id"]
 
-        data = get_batch_export_runs_ok(client, team.pk, batch_export_id)
-        assert len(data["results"]) == 1
-        run = data["results"][0]
-        assert run["status"] == "Running"
+            start_at = "2023-10-23T00:00:00+00:00"
+            end_at = "2023-10-24T00:00:00+00:00"
+            # ensure there is data to backfill, otherwise validation will fail
+            _create_event(
+                team=team,
+                event="$pageview",
+                distinct_id="person_1",
+                timestamp=dt.datetime(2023, 10, 23, 0, 1, 0, tzinfo=dt.UTC),
+            )
+            backfill_batch_export_ok(client, team.pk, batch_export_id, start_at, end_at)
 
-        data = cancel_batch_export_run_ok(client, team.pk, batch_export_id, run["id"])
-        assert data["cancelled"] is True
+            # In order for a run to be cancelable we need a running workflow execution
+            _ = wait_for_workflow_executions(temporal, query=f'TemporalScheduledById="{batch_export_id}"')
 
-        data = get_batch_export_runs_ok(client, team.pk, batch_export_id)
-        assert len(data["results"]) == 1
-        run = data["results"][0]
-        assert run["status"] == "Cancelled"
+            data = get_batch_export_runs_ok(client, team.pk, batch_export_id)
+            assert len(data["results"]) == 1
+            run = data["results"][0]
+            assert run["status"] == "Running"
+
+            data = cancel_batch_export_run_ok(client, team.pk, batch_export_id, run["id"])
+            assert data["cancelled"] is True
+
+            data = get_batch_export_runs_ok(client, team.pk, batch_export_id)
+            assert len(data["results"]) == 1
+            run = data["results"][0]
+            assert run["status"] == "Cancelled"
 
 
 # TODO - add a test to ensure we can't cancel a completed run?
